pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = ""
authors = [
    {name = "bryxosmmm",email = "bryxosmmm@proton.me"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.120.0,<0.121.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "envclasses (>=0.3.1,<0.4.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
